+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master
 
+- Add `broadcastKey` to derive the broadcast endpoint authorization key from the application secret.
+
 ## 0.2.2 (2026-02-12)
 
 - Add missing `/index.js`.